+++ internal/istio/istiofeature/reconcile_canary_operator.go
@@ -80,7 +80,7 @@ func (m *MeshReconciler) installCanaryOperator(c cluster.CommonCluster, promethe
 		values.Operator.Image.Tag = canaryChart.Values.Operator.Image.Tag
 	}
 
-	valuesOverride, err := ConvertStructure(values)
+	valuesOverride, err := helm.ConvertStructure(values)
 	if err != nil {
 		return errors.WrapIf(err, "could not marshal chart value overrides")
 	}
